re-use common code
